Too heck with the 2nd file. Returning the embed link.
